:recycle: move `format_dt` to public
