+++ discord/utils/__init__.py
@@ -50,7 +50,16 @@ from typing import (
 )
 
 from ..errors import HTTPException
-from .public import basic_autocomplete, generate_snowflake, utcnow, snowflake_time, oauth_url, Undefined, MISSING
+from .public import (
+    basic_autocomplete,
+    generate_snowflake,
+    utcnow,
+    snowflake_time,
+    oauth_url,
+    Undefined,
+    MISSING,
+    format_dt,
+)
 
 try:
     import msgspec
@@ -603,53 +612,3 @@ def as_chunks(iterator: _Iter[T], max_size: int) -> _Iter[list[T]]:
     if isinstance(iterator, AsyncIterator):
         return _achunk(iterator, max_size)
     return _chunk(iterator, max_size)
-
-
-TimestampStyle = Literal["f", "F", "d", "D", "t", "T", "R"]
-
-
-def format_dt(dt: datetime.datetime | datetime.time, /, style: TimestampStyle | None = None) -> str:
-    """A helper function to format a :class:`datetime.datetime` for presentation within Discord.
-
-    This allows for a locale-independent way of presenting data using Discord specific Markdown.
-
-    +-------------+----------------------------+-----------------+
-    |    Style    |       Example Output       |   Description   |
-    +=============+============================+=================+
-    | t           | 22:57                      | Short Time      |
-    +-------------+----------------------------+-----------------+
-    | T           | 22:57:58                   | Long Time       |
-    +-------------+----------------------------+-----------------+
-    | d           | 17/05/2016                 | Short Date      |
-    +-------------+----------------------------+-----------------+
-    | D           | 17 May 2016                | Long Date       |
-    +-------------+----------------------------+-----------------+
-    | f (default) | 17 May 2016 22:57          | Short Date Time |
-    +-------------+----------------------------+-----------------+
-    | F           | Tuesday, 17 May 2016 22:57 | Long Date Time  |
-    +-------------+----------------------------+-----------------+
-    | R           | 5 years ago                | Relative Time   |
-    +-------------+----------------------------+-----------------+
-
-    Note that the exact output depends on the user's locale setting in the client. The example output
-    presented is using the ``en-GB`` locale.
-
-    .. versionadded:: 2.0
-
-    Parameters
-    ----------
-    dt: Union[:class:`datetime.datetime`, :class:`datetime.time`]
-        The datetime to format.
-    style: :class:`str`R
-        The style to format the datetime with.
-
-    Returns
-    -------
-    :class:`str`
-        The formatted string.
-    """
-    if isinstance(dt, datetime.time):
-        dt = datetime.datetime.combine(datetime.datetime.now(), dt)
-    if style is None:
-        return f"<t:{int(dt.timestamp())}>"
-    return f"<t:{int(dt.timestamp())}:{style}>"
